csapp:labcache: Complete part B's first matrix
I decided to move on with this lab without finishing the last two
matrices. I've got a good understanding of how cache works. However, to
complete these kinds of assignments demands much of a "hacky" mindset
and ingenuity that I lack. I spent a good number of hours on this lab
but hit a big wall trying to get the full score with the first
matrix. I needed to get some hints from other students to get the idea
of using temp variables to reduce conflict misses. The solution is not
something I find beautiful and elegant so I do not find much
satisfaction solving these kinds of problems.