dis/lab4b: End lab 4b
I decided to stop here because I think this stuff is somewhat just too
difficult for me. I discovered a bug in the snapshot implementation but
since I had spent a significant amount of time on the project, I thought
it'd be better for me to move on and revisit distributed systems in
another course. It might just be too hard without any help from TA or
being able to submit past work. In addition, the later readings from the
course were way over my head.